move the test suite
