+++ pontoon/sync/tests/test_changeset.py
@@ -1,4 +1,3 @@
-from pathlib import Path
 from unittest.mock import Mock, MagicMock, patch
 
 import pytest
@@ -480,53 +479,6 @@ class ChangeSetTests(FakeCheckoutTestCase):
             self.changeset.execute_update_vcs()
             assert mock_has_changed.called
             assert len(resource_file.save.mock_calls) == 0
-
-
-class ChangedConfigFilesTests(FakeCheckoutTestCase):
-    """
-    Tests the algorithm that detects changes of Project Config files.
-    """
-
-    def test_no_config_changes(self):
-        changed_source_files = {"file1.po": [], "test.toml": []}
-
-        with patch.object(
-            self.vcs_project, "configuration"
-        ) as changed_config_files_mock, patch.object(
-            self.vcs_project, "changed_source_files", return_value=changed_source_files
-        ) as changed_source_files_mock:
-            changed_config_files_mock.parsed_configuration.configs.__iter__.return_value = (
-                set()
-            )
-            changed_source_files_mock.__getitem__.return_value = changed_source_files
-            self.assertSetEqual(self.vcs_project.changed_config_files, set())
-
-    def test_changed_config_files(self):
-        config_file_mock = MagicMock()
-        config_file_mock.path = str(
-            Path(self.vcs_project.source_directory_path).joinpath(
-                Path("test-l10n.toml")
-            )
-        )
-        changed_config_files = [config_file_mock]
-        changed_source_files = {
-            "file1.po": [],
-            "test-l10n.toml": [],
-        }
-
-        with patch.object(
-            self.vcs_project, "configuration"
-        ) as changed_config_files_mock, patch.object(
-            self.vcs_project, "changed_source_files", return_value=changed_source_files
-        ) as changed_source_files_mock:
-            changed_config_files_mock.parsed_configuration.configs.__iter__.return_value = (
-                changed_config_files
-            )
-            changed_source_files_mock.__getitem__.return_value = changed_source_files
-
-            self.assertSetEqual(
-                self.vcs_project.changed_config_files, {"test-l10n.toml"}
-            )
 
 
 class AuthorsTests(FakeCheckoutTestCase):
